requirements.txt
pandas
kagglehub[pandas-datasets]
google-cloud-bigquery
google-cloud-storage
pyarrow
